cleanup: kill intgen_t
intgen_t is for jdm (libhandle) interfaces, not as a random
substitute for "int". Do a global s/intgen_t/int across the codebase
with sed to remove this stupidity so that we don't have to include
xfs/jdm.h across almost the entire code base.

After a manual pass back over the callers of the jdm API, there are
no places where the return value needs to be an intgen_t. Indeed,
jdm_open passes back a file descriptor, so intgen_t by definition
cannot be anything other than an int.

Signed-off-by: Dave Chinner <dchinner@redhat.com>