cleanup: move fold_t out of util.h
Many files just want the fold_t type, but pulling in util.h requires
all the XFS and JDM headers to be pulled in, too. Move fold_t to
mlog.[ch] as it tends to be common with log output. Similarly, move
the PREEMPT_* definitions to common/types.h to remove that
dependency on util.h as well.

Signed-off-by: Dave Chinner <dchinner@redhat.com>